Finish Sections
I added a featured section that contains five cards with  a header, an icon, and and a paragraph. I used grid po position the elements
of the cards and flexbox to position the cards itself. Also, I made a line for the headers of the sections using div and some css tricks.